ISSUE-176 Relocate SimpleSessionProvider from calendar-rest-api to storage-api
